the Graph railway is now created properly I think

+++ src/Network.h
@@ -14,23 +14,23 @@ using namespace std;
 class Network : public Graph, public Edge {
 public:
     Network(string stationA, string stationB);
-    Network(string stationA, string stationB, int capacity, string service);
+    Network(string stationA, string stationB, double capacity, string service);
     bool operator<(const Network& network2) const;
 
     const string getStationA() const;
     const string getStationB() const;
-    const int getCapacity() const;
+    const double getCapacity() const;
     const string getService() const;
 
     void setStationA(const string stationA);
     void setStationB(const string stationB);
-    void setCapacity(const int capacity);
+    void setCapacity(const double capacity);
     void setService(const string service);
 
 private:
     string stationA_;
     string stationB_;
-    int capacity_;
+    double capacity_;
     string service_;
 };
